validation for blockspace capacity

+++ world-chain-builder/src/node/args.rs
@@ -1,3 +1,4 @@
+use clap::value_parser;
 use reth_node_optimism::args::RollupArgs;
 
 /// Parameters for rollup configuration
@@ -27,6 +28,6 @@ pub struct WorldChainBuilderArgs {
     /// verified transactions to fill the capacity, the remaining blockspace will be filled with
     /// unverified transactions.
     /// This arg is a percentage of the total blockspace with the default set to 70 (ie 70%).
-    #[arg(long = "builder.verified_blockspace_capacity", default_value = "70")]
-    pub verified_blockspace_capacity: u64,
+    #[arg(long = "builder.verified_blockspace_capacity", default_value = "70", value_parser = value_parser!(u8).range(0..=100))]
+    pub verified_blockspace_capacity: u8,
 }
